// frontend/src/backend.d.ts
export interface PayoutRecord {
    id: string;
    payoutMethod: string;
    destinationAccount: string;
    artistId: string;
    currency: string;
    notes?: string;
    associatedMediaId?: string;
    timestamp: bigint;
    amount: bigint;
}
// ...
export interface InvoiceRecord {
    id: string;
    artistId: string;
    description: string;
    currency: string;
    notes?: string;
    timestamp: bigint;
    amount: bigint;
    mediaId?: string;
}
// ...
export interface MembershipFeeRecord {
    id: string;
    artistId: string;
    currency: string;
    notes?: string;
    timestamp: bigint;
    amount: bigint;
}